requirements.txt
spotipy==2.25.0
python-dotenv==1.0.0
